Get FHiCL parameter as unsigned

+++ artdaq/Generators/CommandableFragmentGenerator.cc
@@ -81,7 +81,7 @@ artdaq::CommandableFragmentGenerator::CommandableFragmentGenerator(const fhicl::
 		fragment_ids.emplace_back(fragment_id);
 	}
 
-	auto generated_fragments_per_event = ps.get<int>("generated_fragments_per_event", 1);
+	auto generated_fragments_per_event = ps.get<size_t>("generated_fragments_per_event", 1);
 	if (generated_fragments_per_event != fragment_ids.size()) {
 		latest_exception_report_ = R"(Error in CommandableFragmentGenerator: "generated_fragments_per_event" disagrees with size of "fragment_ids" list!)";
 		TLOG(TLVL_ERROR) << latest_exception_report_;
